(These are time it took for the chinese + english subtitles to load)
**Unoptimized Version**

32 Seconds - no subtitle video

34 seconds - Hong Kong subtitle video

6 seconds - Chinese subtitle video

**Optimized version**

6 seconds - No subtitle video (433% increase in speed)

7 seconds - Hong Kong subtitle video (386% increase in speed)

6 seconds - Chinese subtitle video